Fixed RNG to not clone stupidly
Originally, it had Copy, which would mean each time I passed in an Rng somewhere, it would use the same seed, so something like StatisticsRoll would end up with a million of the same roll. Now, the seed it wrapped in an Rc<Mutex> to share a mutable seed when cloned.